app/requirements.txt
sklearn==0.0
SQLAlchemy==1.4.22
django-dotenv==1.4.2
python-dotenv==0.19.0
psycopg2-binary
celery==4.4.2
ipdb==0.13.9
redis==3.5.3